+++ lib/koala/http_service/multipart_request.rb
@@ -3,7 +3,7 @@ require 'faraday'
 module Koala  
   module HTTPService
     class MultipartRequest < Faraday::Request::Multipart
-      # Facebook expects nested parameters to be passed in a certain way
+      # Passport expects nested parameters to be passed in a certain way
       # Based on our testing (https://github.com/arsduo/koala/issues/125),
       # Faraday needs two changes to make that work:
       # 1) [] need to be escaped (e.g. params[foo]=bar ==> params%5Bfoo%5D=bar)